requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP & Scraping
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# Database
supabase>=2.0.0

# Configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging
loguru>=0.7.0

# Utilities
python-dateutil>=2.8.0
